+++ src/main/java/racing/domain/NumberGenerator.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain;
 
 public interface NumberGenerator {
