+++ src/main/java/racing/dto/CarDTO.java
@@ -1,4 +1,6 @@
-package racing;
+package racing.dto;
+
+import racing.domain.Car;
 
 public class CarDTO {
