+++ src/test/java/racing/RacingTest.java
@@ -1,9 +1,12 @@
 package racing;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racing.domain.NumberGenerator;
+import racing.domain.Racing;
+import racing.domain.RandomNumberGenerator;
+import racing.dto.RoundResult;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
